requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-JWT-Extended==4.6.0
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pip==24.0
PyJWT==2.8.0
setuptools==41.2.0
Werkzeug==3.0.3
zipp==3.18.2
mysql-connector-python==8.0.33
